Add testsuite
The idea is to run the POSIX filesystem testsuite against a loopback
mounted filesystem and then rerun it once again over a ciopfs mountpoint.

In theory the resulting ${fs}.result and ciopfs-${fs}.result files should
be identical. In practice some tests fail due to timing issues, but it
is nonetheless useful to get an idea on what is working and what isn't.

Run with something like:

 CIOPFS="/path/to/ciopfs-binary-to-test" ./testsuite.sh some-dir

Signed-off-by: Marc Andre Tanner <mat@brain-dump.org>